Civic_Connect-main/existing_yolo_modules/requirements.txt
streamlit
ultralytics
torch
torchvision
Pillow
opencv-python-headless
